Added in progress tracking and reward scoring to the Eye Contact Game.
Former-commit-id: d968f7b5761388f376357a99cb22c46bf84b8bcd

+++ app/src/main/java/edu/odu/cs411/loqui/FirestoreWorker.java
@@ -94,7 +94,7 @@ public class FirestoreWorker
                 });
     }
 
-    public void addToRewardScore(int amount)
+    public void addToRewardScore(double amount)
     {
         DocumentReference userRef = db.collection("users").document(userID);
